Nettools refactor
 - [x] sriov code moved to a separate file
 - [x] calico code moved to a separate file
 - [x] tap code moved to correct file
 - [x] SetupContainerSideNetwork split into smaller parts
